requirements.txt
asgiref==3.5.2
astroid==2.12.12
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dill==0.3.6
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
flake8==5.0.4
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mysqlclient==2.1.1
oauthlib==3.2.2
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pydocstringformatter==0.7.2
pyflakes==2.5.0
PyJWT==2.6.0
pylint==2.15.5
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
